scripts/importer/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
google-generativeai==0.5.2
